fix(--):Solve the problem of large-scale forks on October 1         1, optimize red card verification logic         2, strict inspection in advance of the block         3, solve the problem of rollback failure         4, optimize the block logic, reduce the possibility of fork         5, atomize some functions to avoid abnormal conditions and incomplete processing

+++ consensus-module/poc/consensus-poc-protocol/src/main/java/io/nuls/consensus/poc/protocol/entity/RedPunishData.java
@@ -76,7 +76,7 @@ public class RedPunishData extends TransactionLogicData {
         this.address = address;
     }
 
-    public short getReasonCode() {
+    public byte getReasonCode() {
         return reasonCode;
     }
